added check for game version

+++ game/main.py
@@ -22,6 +22,7 @@ from testing.TestingHandler import TestingHandler
 
 # IO game packages
 from IO import Window
+from IO.Input import enter_to_continue
 
 # built-in packages
 import sys
@@ -36,7 +37,7 @@ import argparse
 console = Console()  # the console
 Window.clear()  # clear window
 
-version = "V1.0.0"
+version = "V1.1.0"
 
 parser = argparse.ArgumentParser(
     prog="Gentry's Quest",
@@ -70,19 +71,31 @@ else:
     if args.server is None:
         server = Server("https://gdcheerios.com")  # default server url
     else:
-        server = Server(args.server)  # make cl ass to store server info
+        server = Server(args.server)  # make class to store server info
     if args.username is not None and args.password is not None:
+        latest_version = server.API.get_version()
+        version_differs = version != latest_version
         account_info = AccountInfo(args.username, args.password)  # make class to store account info
         user_data = server.API.login(account_info.username, account_info.password)  # game data class initialization
         user = User(user_data["id"], account_info.username, server.API.get_power_level())  # user class initialization
         game_data = GameData(user_data["metadata"]["Gentry's Quest data"])
         game = Game(game_data, version, server)
+        if version_differs:
+            WarningText("You are not on the right Gentry's Quest version!\n"
+                        f"Your version: {version}\n"
+                        f"latest: {latest_version}\n"
+                        "You will not be able to use online features...\n"
+                        "You can get the latest version from [link=https://gdcheerios.com/gentrys-quest]https://gdcheerios.com/gentrys-quest[/link].").display()
+            enter_to_continue()
+            server.API.check_out()
+            server.disable()
         def byebye():
             game_status = Status("Uploading data...")
             game_status.start()
-            server.API.upload_data(game.game_data)
-            server.API.check_out()
-            server.API.token.delete()
+            if not server.disabled:
+                server.API.upload_data(game.game_data)
+                server.API.check_out()
+                server.API.token.delete()
             game_status.stop()
 
         atexit.register(byebye)
